chore(env): rename environment.test.ts to environment.ts
