+++ frontend/src/app/services/auth.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { map, Observable, tap } from 'rxjs';
-import { environment } from '../../environments/environment.test';
+import { environment } from '../../environments/environment';
 
 @Injectable({
   providedIn: 'root',
